Still doing more anim stuff + Bug Fixes
Fixed the bug that caused the logic for facedirection variable to not be the same like in the other parts of the script. Turns out that the player was just not oriented correctly and a whole mess was made. Fixed tho, hopefully it makes things easier in the future to understand.